yby_withcoaches updated
The year_by_year_to2015_withcoaches file now has values for the Coach
and Opponent_Coach columns for all games coached by a Coach who coached
at some point between 2005 and 2015.  While there should be no NAs in
the Coach column there are some in the Opponent_Coach column where the
Opponent was an FCS team that we don't have data for.